Remove BillingHelpers, move methods to bare model
